+++ src/stories/basic.css
@@ -8,7 +8,7 @@ canvas {
 
 .info {
   color: #FFF;
-  position: absolute;
+  position: absolute; top:0; left: 0;
   padding: 20px 0 0 20px;
 }
